floor function

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 '''a type-annotated function floor which takes a float n as argument'''
+
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     '''returns the floor of the float.'''
     return math.floor(n)
